change moudle structure

+++ src/main/java/com/codingdie/analyzer/spider/model/PageTask.java
@@ -14,7 +14,7 @@ public class PageTask implements Serializable{
     public PageTask(int pn){
         this.pn=pn;
     }
-    public int pn=50;
+    public long pn=50;
     public int status=STATUS_TODO;
 
 }
